Fix extraction of encoding from meta element
When a meta element was parsed, with a http-equiv of `content-type` it
was using `mimetools` to do so.  This was deprecated since Python 2.3,
and as is pointed out by https://github.com/moinwiki/moin/issues/1089
does not work in Python 3.

This change uses the cgi module, and parse_header in particular to parse
the value correctly.